shapes_grid/Grid_Statistics_AddTo_Polygon: fix, removed vc compiler warning

+++ saga-gis/src/tools/shapes/shapes_grid/Grid_Statistics_AddTo_Polygon.cpp
@@ -551,7 +551,7 @@ bool CGrid_Statistics_AddTo_Polygon::Get_Simple_Index(CSG_Shapes *pPolygons, CSG
 
 					if( bFill )
 					{
-						Index.Set_Value(x, y, iShape);
+						Index.Set_Value(x, y, (double)iShape);
 					}
 				}
 			}
